test(openai): remove unsupported live OAuth probes
Punchcard-Session: cobalt-cedar-timber-04
Punchcard-Session: calm-cedar-river-aa
Punchcard-Session: golden-meadow-cedar-dv
