+++ extensions/openai/realtime-quicksilver-session.ts
@@ -3,7 +3,6 @@ import { randomBytes, randomUUID } from "node:crypto";
 import type { IncomingMessage, ServerResponse } from "node:http";
 import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
 import type { PluginLogger } from "openclaw/plugin-sdk/plugin-entry";
-import { resolveProviderAuthProfileApiKey } from "openclaw/plugin-sdk/provider-auth";
 import type {
   RealtimeVoiceBridge,
   RealtimeVoiceBrowserSession,
@@ -15,7 +14,6 @@ import {
   resolveAcceptedBrowserOrigin,
 } from "openclaw/plugin-sdk/webhook-request-guards";
 import WebSocket, { type RawData } from "ws";
-import { resolveCodexAuthIdentity } from "./openai-chatgpt-auth-identity.js";
 import { OpenAIQuicksilverDelegationController } from "./realtime-quicksilver-delegation-controller.js";
 import {
   releaseOpenAIQuicksilverSession,
@@ -148,27 +146,6 @@ function applyRealtimeOfferCorsHeaders(
 function readBearerToken(req: IncomingMessage): string | undefined {
   const authorization = req.headers.authorization?.trim();
   return authorization?.match(/^Bearer\s+([^\s]+)$/i)?.[1];
-}
-
-export async function resolveOpenAIChatGptSubscriptionAuth(params: {
-  cfg?: OpenClawConfig;
-  agentDir?: string;
-}): Promise<Extract<OpenAIQuicksilverAuth, { type: "oauth" }> | undefined> {
-  const token = await resolveProviderAuthProfileApiKey({
-    provider: "openai",
-    cfg: params.cfg,
-    agentDir: params.agentDir,
-    profileTypes: ["oauth"],
-    includeExternalCliAuth: false,
-  });
-  if (!token) {
-    return undefined;
-  }
-  const accountId = resolveCodexAuthIdentity({ accessToken: token }).accountId;
-  if (!accountId) {
-    throw new Error("The selected ChatGPT OAuth profile is missing its account id");
-  }
-  return { type: "oauth", token, accountId };
 }
 
 export function createOpenAIQuicksilverBrowserSessionBroker(params: {
